Organized the core module
