+++ core/src/it/units/sdm/project/core/board/Stone.java
@@ -1,4 +1,4 @@
-package it.units.sdm.project;
+package it.units.sdm.project.core.board;
 
 import org.jetbrains.annotations.NotNull;
