+++ core/src/it/units/sdm/project/interfaces/Game.java
@@ -1,6 +1,6 @@
 package it.units.sdm.project.interfaces;
 
-import it.units.sdm.project.Player;
+import it.units.sdm.project.core.game.Player;
 import org.jetbrains.annotations.Nullable;
 
 public interface Game {
