+++ core/test/java/PlayerTests.java
@@ -1,5 +1,5 @@
-import it.units.sdm.project.Player;
-import it.units.sdm.project.Stone;
+import it.units.sdm.project.core.game.Player;
+import it.units.sdm.project.core.board.Stone;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
