+++ core/test/java/PositionTests.java
@@ -1,5 +1,5 @@
 import it.units.sdm.project.exceptions.InvalidPositionException;
-import it.units.sdm.project.Position;
+import it.units.sdm.project.core.board.Position;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
