+++ desktop/src/it/units/sdm/project/DesktopLauncher.java
@@ -2,6 +2,7 @@ package it.units.sdm.project;
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
+import it.units.sdm.project.core.game.gui.FreedomGame;
 
 // Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
 public class DesktopLauncher {
