+++ html/src/it/units/sdm/project/client/HtmlLauncher.java
@@ -3,7 +3,7 @@ package it.units.sdm.project.client;
 import com.badlogic.gdx.ApplicationListener;
 import com.badlogic.gdx.backends.gwt.GwtApplication;
 import com.badlogic.gdx.backends.gwt.GwtApplicationConfiguration;
-import it.units.sdm.project.FreedomGame;
+import it.units.sdm.project.core.game.gui.FreedomGame;
 
 public class HtmlLauncher extends GwtApplication {
